Refactor NewRequest Component
